kerenal refactor for lavamoat-node support
